# schemas/api.openapi.json
{
  "components": {
    "schemas": {
...
      "CorpusSnapshotResource": {
        "additionalProperties": false,
        "description": "Atomic typed preparation state with explicit source selections.",
        "properties": {
          "acquisition_companies": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ManifestIssuerResource"
            },
            "title": "Acquisition Companies",
            "type": "array"
          },
          "acquisition_draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcquisitionDraftResource"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/CorpusDocumentResource"
            },
            "title": "Documents",
            "type": "array"
          },
          "manifests": {
            "items": {
              "$ref": "#/components/schemas/ManifestResource"
            },
            "title": "Manifests",
            "type": "array"
          },
          "mode": {
            "enum": [
              "live",
              "canned"
            ],
            "title": "Mode",
            "type": "string"
          },
          "sources": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SourceInventoryResource"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/CorpusStatusResource"
          }
        },
        "required": [
          "mode",
          "status",
          "manifests",
          "documents"
        ],
        "title": "CorpusSnapshotResource",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Top-level typed error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Document Review RAG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/corpus": {
      "get": {
        "description": "Return live corpus, schema, index, manifest, and document state.",
        "operationId": "corpus_snapshot_admin_corpus_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusSnapshotResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Corpus Snapshot",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/corpus/jobs": {
      "get": {
        "description": "Return current corpus job queue and bounded history.",
        "operationId": "corpus_jobs_admin_corpus_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusJobsResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Corpus Jobs",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Queue one safe corpus acquisition, ingest, or indexing operation.",
        "operationId": "enqueue_corpus_admin_corpus_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorpusOperationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusJobResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Enqueue Corpus",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/corpus/jobs/{job_id}/retry": {
      "post": {
        "description": "Retry one known failed corpus job.",
        "operationId": "retry_corpus_admin_corpus_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusJobResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Retry Corpus",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/corpus/sources/deletion-preview": {
      "post": {
        "description": "Inspect exact acquired originals without deleting or changing their selection.",
        "operationId": "source_deletion_preview_admin_corpus_sources_deletion_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceDeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDeletionPreviewResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Source Deletion Preview",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/documents/facets": {
      "get": {
        "description": "Return live registry, issuer, year, language, form, and status facets.",
        "operationId": "document_facets_admin_documents_facets_get",
        "parameters": [
          {
            "in": "query",
            "name": "registry",
            "required": false,
            "schema": {
              "default": "",
              "title": "Registry",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentFacetsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Document Facets",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/documents/{doc_id}": {
      "get": {
        "description": "Return bounded metadata and chunk previews for one document.",
        "operationId": "document_detail_admin_documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Document Detail",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/evaluations/compare": {
      "get": {
        "description": "Return metrics and per-case changes between compatible artifacts.",
        "operationId": "compare_evaluations_admin_evaluations_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Candidate Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "baseline_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Baseline Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Compare Evaluations",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}